requirements.txt
flask
python-binance
python-dotenv
requests
pandas
numpy
gspread
google-auth
pyTelegramBotAPI